Pt_5_2: adding and searching in a csv file
the user enters how many entries he wants to add to the list. The program adds data through a loop to the dictionary, which is added to the dataframe. A csv file is created from the received dataframe.
Then the user enters the author. The search is performed by the column 'Author' and all lines with the entered author are displayed, otherwise 'Автор не найден'